pyproject.toml
[tool.poetry]
name = "garticmon-bot"
version = "0.1.0"
description = ""
authors = ["Steven B <373152+Stevebel@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
discord = "^2.3.0"
python-dotenv = "^1.0.0"
unidecode = "^1.3.6"
Pillow = "^9.0.0"
pypng = "0.0.21"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
